When lyric time > 65535, lyric overlapped. fixed; don't bold lyric when time is 0.
--HG--
extra : convert_revision : svn%3A52058a13-6432-4f25-a4c3-f644898cbfcb%4068

+++ src/foo_alsong_lyric/Common_Lyric_Manipulation.h
@@ -7,7 +7,7 @@ private:
 	string Album;
 	string Artist;
 	string Registrant;
-	vector<WORD> Time;
+	vector<DWORD> Time;
 
 	static DWORD GetFileHash(unsigned char *Data, int Size, CHAR *Hash, CHAR *fmt);
 	DWORD ParseLyric(CHAR *InputLyric, CHAR *Delimiter);
